chore(integer): move u256 into its own mod
