+++ tfhe/src/integer/client_key/mod.rs
@@ -19,6 +19,7 @@ use crate::shortint::{
 use serde::{Deserialize, Serialize};
 pub use utils::radix_decomposition;
 
+use crate::integer::U256;
 pub use crt::CrtClientKey;
 pub use radix::RadixClientKey;
 
@@ -79,30 +80,6 @@ impl ClearText for u128 {
         unsafe { std::slice::from_raw_parts_mut(u128_slc.as_mut_ptr() as *mut u64, 2) }
     }
 }
-#[derive(Default, Copy, Clone, Debug, PartialEq, Eq)]
-pub struct U256([u128; 2]);
-
-impl U256 {
-    #[inline]
-    pub fn low(&self) -> u128 {
-        self.0[0]
-    }
-
-    #[inline]
-    fn high(&self) -> u128 {
-        self.0[1]
-    }
-
-    #[inline]
-    pub fn low_mut(&mut self) -> &mut u128 {
-        &mut self.0[0]
-    }
-
-    #[inline]
-    pub fn high_mut(&mut self) -> &mut u128 {
-        &mut self.0[1]
-    }
-}
 
 impl ClearText for U256 {
     fn as_words(&self) -> &[u64] {
@@ -113,66 +90,6 @@ impl ClearText for U256 {
     fn as_words_mut(&mut self) -> &mut [u64] {
         let u128_slc = self.0.as_mut_slice();
         unsafe { std::slice::from_raw_parts_mut(u128_slc.as_mut_ptr() as *mut u64, 4) }
-    }
-}
-
-#[cfg(test)]
-impl rand::distributions::Distribution<U256> for rand::distributions::Standard {
-    fn sample<R: rand::Rng + ?Sized>(&self, rng: &mut R) -> U256 {
-        let low = rng.gen::<u128>();
-        let high = rng.gen::<u128>();
-        U256::from((low, high))
-    }
-}
-
-// Since we store as [low, high], deriving ord
-// would produces bad ordering
-impl std::cmp::Ord for U256 {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        let high_bits_ord = self.high().cmp(&other.high());
-        if let std::cmp::Ordering::Equal = high_bits_ord {
-            self.low().cmp(&other.low())
-        } else {
-            high_bits_ord
-        }
-    }
-}
-
-impl std::ops::Add<Self> for U256 {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        let (new_low, has_overflowed) = self.low().overflowing_add(rhs.low());
-        let new_high = self
-            .high()
-            .wrapping_add(rhs.high())
-            .wrapping_add(u128::from(has_overflowed));
-
-        Self::from((new_low, new_high))
-    }
-}
-
-impl std::ops::AddAssign<Self> for U256 {
-    fn add_assign(&mut self, rhs: Self) {
-        *self = *self + rhs;
-    }
-}
-
-impl std::cmp::PartialOrd for U256 {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl From<(u128, u128)> for U256 {
-    fn from(v: (u128, u128)) -> Self {
-        Self([v.0, v.1])
-    }
-}
-
-impl From<u128> for U256 {
-    fn from(value: u128) -> Self {
-        Self::from((value, 0))
     }
 }
